Refactors BraintreePayPal to inherit from PayPal
- Moves failure and error handlers to PayPal class
- Moves PayPal factory to separate module to avoid circular reference
- Adds non-factory instantiation error detection
