+++ lib/errors.js
@@ -209,7 +209,7 @@ errors.add('apple-pay-config-invalid', {
 });
 
 errors.add('apple-pay-factory-only', {
-  message: 'Apple Pay must be initialized by calling recurly.applePay'
+  message: 'Apple Pay must be initialized by calling recurly.ApplePay'
 });
 
 errors.add('apple-pay-init-error', {
@@ -222,6 +222,10 @@ errors.add('apple-pay-init-error', {
 
 errors.add('apple-pay-payment-failure', {
   message: 'Apply Pay could not charge the customer'
+});
+
+errors.add('paypal-factory-only', {
+  message: 'PayPal must be initialized by calling recurly.PayPal'
 });
 
 errors.add('paypal-config-missing', {
