+++ lib/recurly.js
@@ -12,7 +12,8 @@ import errors from './errors';
 import version from './version';
 import bankAccount from './recurly/bank-account';
 import {factory as ApplePay} from './recurly/apple-pay';
-import {factory as PayPal, deprecatedPaypal} from './recurly/paypal';
+import {factory as PayPal} from './recurly/paypal/factory';
+import {deprecatedPaypal} from './recurly/paypal';
 import {Bus} from './recurly/bus';
 import {Fraud} from './recurly/fraud';
 import {HostedFields, FIELD_TYPES} from './recurly/hosted-fields';
