+++ lib/recurly/paypal/braintree.js
@@ -1,35 +1,25 @@
-import Emitter from 'component-emitter';
+import omit from 'lodash.omit';
 import after from 'lodash.after';
 import loadScript from 'load-script';
-import omit from 'lodash.omit';
+import Emitter from 'component-emitter';
 import {PayPal} from './';
 import errors from '../../errors';
 
 const debug = require('debug')('recurly:paypal:braintree');
 
-const BRAINTREE_CLIENT_VERSION = '3.10.0';
+const BRAINTREE_CLIENT_VERSION = '3.11.0';
 
 /**
  * Braintree-specific PayPal handler
- *
- * TODO: make inherit from PayPal instead of Emitter to consolidate error handler and init
  */
 
-export class BraintreePayPal extends Emitter {
-  constructor (options) {
-    super();
-    this.ready = false;
-    this.config = {};
-    this.configure(options);
-  }
-
+export class BraintreePayPal extends PayPal {
   configure (options) {
+    super.configure(options);
     if (!options.braintree || !options.braintree.clientAuthorization) {
-      throw this.error('paypal-config-missing', { opt: 'braintree.clientAuthorization'})
+      return this.error('paypal-config-missing', { opt: 'braintree.clientAuthorization'})
     }
     this.config.clientAuthorization = options.braintree.clientAuthorization;
-    this.recurly = options.recurly;
-    this.constructorOptions = options;
     this.load();
   }
 
@@ -58,9 +48,8 @@ export class BraintreePayPal extends Emitter {
     braintree.client.create({ authorization }, (error, client) => {
       if (error) return this.fail('paypal-braintree-api-error', { error });
       debug('Braintree client created');
-
       braintree.dataCollector.create({ client, paypal: true }, (error, collector) => {
-        if (error) return this.error('paypal-braintree-api-error', { error });
+        if (error) return this.fail('paypal-braintree-api-error', { error });
         debug('Device data collector created');
         this.deviceFingerprint = collector.deviceData;
         braintree.paypal.create({ client }, (error, paypal) => {
@@ -115,21 +104,10 @@ export class BraintreePayPal extends Emitter {
    */
   fail (reason, options) {
     debug('Failure scenario encountered. Falling back to direct PayPal flow', reason, options);
-    if (reason) this.error(reason, options);
-    return new PayPal(omit(this.constructorOptions, 'braintree'));
-  }
-
-  /**
-   * Creates and emits a RecurlyError
-   *
-   * @param  {...Mixed} params to be passed to the Recurlyerror factory
-   * @return {RecurlyError}
-   * @emit 'error'
-   * @private
-   */
-  error (...params) {
-    let err = params[0] instanceof Error ? params[0] : errors(...params);
-    this.emit('error', err);
-    return err;
+    let failure = this.failure = {};
+    if (reason) failure.error = this.error(reason, options);
+    failure.fallback = new PayPal(omit(this.config.constructorOptions, 'braintree'));
+    this.emit('failure', failure);
+    return failure;
   }
 }
