+++ lib/recurly/paypal/index.js
@@ -1,25 +1,20 @@
 import Emitter from 'component-emitter';
-import {BraintreePayPal} from './braintree';
+import {Recurly} from '../../recurly';
 
 const debug = require('debug')('recurly:paypal');
-
-/**
- * Instantiation factory
- *
- * @param {Object} options
- * @param {Function} done
- * @return {PayPal}
- */
-export function factory (options, done) {
-  options = Object.assign({}, options, { recurly: this });
-  if (options.braintree) return new BraintreePayPal(options);
-  return new PayPal(options);
-};
 
 export class PayPal extends Emitter {
   constructor (options) {
     super();
+    this.ready = false;
+    this.config = {};
+    this.configure(options);
+  }
+
+  configure (options) {
+    if (!(options.recurly instanceof Recurly)) return this.error('paypal-factory-only');
     this.recurly = options.recurly;
+    this.config.constructorOptions = options;
   }
 
   /**
@@ -58,4 +53,4 @@ export class PayPal extends Emitter {
 export function deprecatedPaypal (data, done) {
   debug('start');
   this.open('/paypal/start', data, done);
-};
+}
